putting everything together
