+++ tests/test-full-producer.cpp
@@ -47,7 +47,8 @@ BOOST_AUTO_TEST_CASE(OnInterest)
   Name syncInterestName(syncPrefix);
   syncInterestName.append("malicious-IBF");
 
-  BOOST_REQUIRE_NO_THROW(node.onSyncInterest(syncPrefix, Interest(syncInterestName)));
+  BOOST_REQUIRE_NO_THROW(node.m_producerArbitrary.onSyncInterest(syncPrefix,
+  	                                                             Interest(syncInterestName)));
 }
 
 BOOST_AUTO_TEST_SUITE_END()
